Add q3_K KV cache type + fork README
Adds q3_K (3.44-bit K-quant) as a selectable KV cache type for the CUDA
backend: --cache-type-k q3_K / --cache-type-v q3_K. ~24% less KV-cache
VRAM than q4_0, lossless on large models (Qwen 27B perplexity unchanged
within noise), and full prefill speed (~940 tok/s pp4096 on RTX 3090).

Wires q3_K into the KV-cache paths: cache-type parsing (arg.cpp), the
SET_ROWS quantized cache write (cpy-utils.cuh, set-rows.cu, ggml-cuda.cu),
the Flash-Attention support gate and forced f16 conversion (fattn.cu),
and a non-contiguous q3_K->f16 dequant for the strided KV view
(convert.cu). llama-bench also accepts -ctk/-ctv q3_K. Plus a uint32
narrowing fix in llama-model.cpp so the CUDA build passes on MSVC.

Replaces the README with a fork-focused one describing both changes
(this q3_K type and the #27109 prefill fix); the original upstream
README is preserved as README.upstream.md.

+++ src/llama-model.cpp
@@ -542,7 +542,7 @@ struct ggml_backend_meta_split_state llama_meta_device_get_split_state(const str
                     return {{key_dim, 2}, {value_dim, 1}};
                 }
             } else {
-                const int64_t head_ratio = n_v_heads / n_k_heads;
+                const uint32_t head_ratio = (uint32_t)(n_v_heads / n_k_heads);
                 if (std::regex_match(tensor_name, pattern_qkv_weight) || std::regex_match(tensor_name, pattern_ssm_conv1d)) {
                     GGML_ASSERT(tensor->ne[axis] == 2*key_dim + value_dim);
                     return {{key_dim, 2 + head_ratio}};
